Group common fixtures of conftest.py
