+++ noxfile.py
@@ -5,7 +5,8 @@ import nox
 
 @nox.session
 # @nox.parametrize("borgbackup", ["2.0.0b5", "2.0.0b4"])
-@nox.parametrize("borgbackup", ["1.1.18", "1.2.2", "1.2.3", "1.2.4"])
+# @nox.parametrize("borgbackup", ["1.1.18", "1.2.2", "1.2.3", "1.2.4"])
+@nox.parametrize("borgbackup", ["1.2.4"])
 def run_tests(session, borgbackup):
     # install borgbackup
     session.install(f"borgbackup[pyfuse3]=={borgbackup}")
