+++ tests/integration/conftest.py
@@ -1,6 +1,5 @@
 import os
 import subprocess
-import sys
 import time
 
 import pytest
@@ -32,34 +31,6 @@ models = [
     EventLogModel,
     SchemaVersion,
 ]
-
-
-def pytest_configure(config):
-    sys._called_from_test = True
-    pytest._wait_defaults = {'timeout': 20000}
-    os.environ['LANG'] = 'en'  # Ensure we test an English UI
-
-
-@pytest.fixture(scope='session')
-def qapp(tmpdir_factory):
-    # DB is required to init QApplication. New DB used for every test.
-    tmp_db = tmpdir_factory.mktemp('Vorta').join('settings.sqlite')
-    mock_db = SqliteDatabase(str(tmp_db))
-    vorta.store.connection.init_db(mock_db)
-
-    # Needs to be disabled before calling VortaApp()
-    if sys.platform == 'darwin':
-        cfg = vorta.store.models.SettingsModel.get(key='check_full_disk_access')
-        cfg.value = False
-        cfg.save()
-
-    from vorta.application import VortaApp
-
-    qapp = VortaApp([])  # Only init QApplication once to avoid segfaults while testing.
-
-    yield qapp
-    mock_db.close()
-    qapp.quit()
 
 
 @pytest.fixture(scope='function', autouse=True)
@@ -201,16 +172,6 @@ def choose_file_dialog(tmpdir):
                 return [str(tmpdir)]
 
     return MockFileDialog
-
-
-@pytest.fixture
-def borg_json_output():
-    def _read_json(subcommand):
-        stdout = open(f'tests/borg_json_output/{subcommand}_stdout.json')
-        stderr = open(f'tests/borg_json_output/{subcommand}_stderr.json')
-        return stdout, stderr
-
-    return _read_json
 
 
 @pytest.fixture
